+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Operator overloading for real APy* and Python complex.
 - ROM generators for (System)Verilog, Spade, Chisel, Amaranth, and Spinal HDL.
 - Getters for real and imaginary part of `APyCFloatArray`.
+- Support for logic NOT, AND, OR, and XOR in `APyCFloat`.
 
 ### Fixed